feat: add support for generic JSON -> ApiDOM transformation
Use generic JSON adapter to base for OpenApi 3.1.x JSON adapter.

Refs https://github.com/swagger-api/oss-planning/issues/133
