+++ apidom/packages/apidom-parser-adapter-json/.eslintignore
@@ -1,0 +1,8 @@
+/dist
+/es
+/lib
+/types
+/config
+/.nyc_output
+/node_modules
+/**/*.js
