+++ apidom/packages/apidom-parser-adapter-json/.gitignore
@@ -1,0 +1,4 @@
+/dist
+/es
+/lib
+/types
