+++ apidom/packages/apidom-parser-adapter-json/.mocharc.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = {
+  recursive: true,
+  spec: 'test/**/*.ts',
+  require: ['test/mocha-bootstrap.js'],
+};
