+++ apidom/packages/apidom-parser-adapter-json/.npmrc
@@ -1,0 +1,2 @@
+save-prefix="="
+save=false
