+++ apidom/packages/apidom-parser-adapter-json/package.json
@@ -1,12 +1,12 @@
 {
-  "name": "apidom-parser-adapter-openapi-json-3-1",
+  "name": "apidom-parser-adapter-json",
   "version": "0.1.0",
-  "description": "OpenAPI 3.1.x JSON adapter for ApiDOM parser",
+  "description": "Generic JSON adapter for ApiDOM parser",
   "main": "lib/adapter-node.js",
   "module": "es/adapter-browser.js",
   "jsnext:main": "es/adapter-browser.js",
   "browser": "es/adapter-browser.js",
-  "unpkg": "dist/apidom-parser-apdater-openapi-json-3-1.browser.min.js",
+  "unpkg": "dist/apidom-parser-apdater-json.browser.min.js",
   "scripts": {
     "build": "npm run build:es && npm run build:commonjs && npm run build:umd:browser",
     "build:es": "cross-env BABEL_ENV=es babel src --out-dir es --extensions '.ts' --root-mode 'upward'",
@@ -27,8 +27,6 @@
     "@apidevtools/json-schema-ref-parser": "=9.0.3",
     "apidom": "file:../apidom",
     "apidom-ast": "file:../apidom-ast",
-    "apidom-ns-openapi-3-1": "file:../apidom-ns-openapi-3-1",
-    "apidom-parser-adapter-json": "file:../apidom-parser-adapter-json",
     "ramda": "=0.27.0",
     "@types/ramda": "=0.27.6",
     "ramda-adjunct": "=2.27.0",
