+++ apidom/packages/apidom-parser-adapter-json/src/adapter-browser.ts
@@ -1,0 +1,21 @@
+export { default as parse, namespace } from './parser/index-browser';
+export { detect, mediaTypes } from './adapter';
+
+export { default as specification } from './parser/specification';
+export { addSourceMap } from './parser/source-map';
+export { appendMetadata } from './parser/metadata';
+export { hasKey, hasKeys } from './parser/predicates';
+
+export { default as Visitor } from './parser/visitors/Visitor';
+export { default as SpecificationVisitor } from './parser/visitors/SpecificationVisitor';
+export { default as ErrorVisitor } from './parser/visitors/ErrorVisitor';
+export { default as DocumentVisitor } from './parser/visitors/DocumentVisitor';
+
+export { ValueVisitor, ArrayVisitor, ObjectVisitor } from './parser/visitors/generics';
+export { default as AlternatingVisitor } from './parser/visitors/generics/AlternatingVisitor';
+export { default as FixedFieldsJsonObjectVisitor } from './parser/visitors/generics/FixedFieldsJsonObjectVisitor';
+export { default as MapJsonObjectVisitor } from './parser/visitors/generics/MapJsonObjectVisitor';
+export { default as MixedFieldsJsonObjectVisitor } from './parser/visitors/generics/MixedFieldsJsonObjectVisitor';
+export { default as PatternedFieldsJsonObjectVisitor } from './parser/visitors/generics/PatternedFieldsJsonObjectVisitor';
+
+export { visit, BREAK } from './parser/visitors';
