+++ apidom/packages/apidom-parser-adapter-json/src/adapter.ts
@@ -1,0 +1,10 @@
+export const mediaTypes = ['application/javascript'];
+
+export const detect = (source: string): boolean => {
+  try {
+    JSON.parse(source);
+  } catch (e) {
+    return false;
+  }
+  return true;
+};
