+++ apidom/packages/apidom-parser-adapter-json/src/parser/metadata.ts
