+++ apidom/packages/apidom-parser-adapter-json/src/parser/visitors/Visitor.ts
@@ -1,4 +1,5 @@
 import stampit from 'stampit';
+
 import { addSourceMap } from '../source-map';
 
 const Visitor = stampit({
