+++ apidom/packages/apidom-parser-adapter-json/src/parser/visitors/generics/MapJsonObjectVisitor.ts
@@ -1,0 +1,12 @@
+import stampit from 'stampit';
+import { isNonEmptyString } from 'ramda-adjunct';
+
+import PatternedFieldsJsonObjectVisitor from './PatternedFieldsJsonObjectVisitor';
+
+const MapJsonObjectVisitor = stampit(PatternedFieldsJsonObjectVisitor, {
+  props: {
+    fieldPatternPredicate: isNonEmptyString,
+  },
+});
+
+export default MapJsonObjectVisitor;
