+++ apidom/packages/apidom-parser-adapter-json/src/polyfills.ts
@@ -1,0 +1,3 @@
+/* eslint-disable import/no-extraneous-dependencies */
+import 'core-js/stable';
+import 'regenerator-runtime/runtime';
