+++ apidom/packages/apidom-parser-adapter-json/test/.eslintrc
@@ -1,0 +1,35 @@
+{
+  "env": {
+    "mocha": true
+  },
+  "globals": {
+    "document": true
+  },
+  "plugins": [
+    "mocha"
+  ],
+  "rules": {
+    "no-void": 0,
+    "func-names": 0,
+    "prefer-arrow-callback": 0,
+    "no-array-constructor": 0,
+    "prefer-rest-params": 0,
+    "no-new-wrappers": 0,
+    "mocha/no-skipped-tests": 2,
+    "mocha/handle-done-callback": 2,
+    "mocha/valid-suite-description": 2,
+    "mocha/no-mocha-arrows": 2,
+    "mocha/no-hooks-for-single-case": 2,
+    "mocha/no-sibling-hooks": 2,
+    "mocha/no-top-level-hooks": 2,
+    "mocha/no-identical-title": 2,
+    "mocha/no-nested-tests": 2,
+    "mocha/no-exclusive-tests": 2
+  },
+  "overrides": [{
+    "files": ["mocha-bootstrap.js"],
+    "parserOptions": {
+      "sourceType": "script"
+    }
+  }]
+}
