+++ apidom/packages/apidom-parser-adapter-json/test/fixtures/sample-data.json
@@ -1,0 +1,3 @@
+{
+  "prop": "val"
+}
