+++ apidom/packages/apidom-parser-adapter-json/test/mocha-bootstrap.js
@@ -1,0 +1,2 @@
+require('regenerator-runtime');
+require('@babel/register')({ extensions: ['.js', '.ts'], rootMode: 'upward' });
