+++ apidom/packages/apidom-parser-adapter-json/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../tsconfig.json",
+  "include": [
+    "../@types/**/*.d.ts",
+    "src/**/*",
+    "test/**/*"
+  ]
+}
