+++ apidom/packages/apidom-parser-adapter-openapi-json-3-1/src/parser/visitors/generics/MixedFieldsJsonObjectVisitor.ts
@@ -1,39 +1,13 @@
 import stampit from 'stampit';
-import { noop } from 'ramda-adjunct';
+// @ts-ignore
+import { MixedFieldsJsonObjectVisitor as BaseVisitor } from 'apidom-parser-adapter-json';
 
-import { BREAK } from '..';
-import FixedFieldsJsonObjectVisitor from './FixedFieldsJsonObjectVisitor';
-import PatternedFieldsJsonObjectVisitor from './PatternedFieldsJsonObjectVisitor';
+import { isOpenApiExtension } from '../../predicates';
 
-const MixedFieldsJsonObjectVisitor = stampit(
-  FixedFieldsJsonObjectVisitor,
-  PatternedFieldsJsonObjectVisitor,
-  {
-    props: {
-      specPathFixedFields: noop,
-      specPathPatternedFields: noop,
-    },
-    methods: {
-      object(objectNode) {
-        const { specPath } = this;
-
-        try {
-          this.specPath = this.specPathFixedFields;
-          // @ts-ignore
-          FixedFieldsJsonObjectVisitor.compose.methods.object.call(this, objectNode);
-
-          this.specPath = this.specPathPatternedFields;
-          // @ts-ignore
-          PatternedFieldsJsonObjectVisitor.compose.methods.object.call(this, objectNode);
-        } catch (e) {
-          this.specPath = specPath;
-          throw e;
-        }
-
-        return BREAK;
-      },
-    },
+const MixedFieldsJsonObjectVisitor = stampit(BaseVisitor, {
+  props: {
+    specificationExtensionPredicate: isOpenApiExtension({}),
   },
-);
+});
 
 export default MixedFieldsJsonObjectVisitor;
