+++ apidom/packages/apidom-parser-adapter-openapi-json-3-1/src/parser/visitors/open-api-3-1/info/VersionVisitor.ts
@@ -1,6 +1,7 @@
 import stampit from 'stampit';
+// @ts-ignore
+import { appendMetadata } from 'apidom-parser-adapter-json';
 
-import { appendMetadata } from '../../../metadata';
 import { ValueVisitor } from '../../generics';
 
 const VersionVisitor = stampit(ValueVisitor, {
